requirements.txt
google-generativeai>=0.5.0
pyperclip
keyboard
pystray
Pillow 